fix: ensure that bridge does not ping python backend
